Add docstrings and edge case tests
