+++ MEMO.md
@@ -1,2 +1,4 @@
 TODOが空だったため、基本的な開発タスクを追加した。
 Python版のビットボードオセロ実装を開始。基本的な合法手計算と着手処理を実装した。
+
+Docstrings added and edge case tests implemented.
