+++ src/main.py
@@ -1,46 +1,6 @@
-import sys
-from othello.board import BitBoard
+"""Entry point for running the command line game."""
 
+from othello.cli import main
 
-def parse_move(move_str: str) -> int:
-    col = ord(move_str[0].lower()) - ord('a')
-    row = int(move_str[1]) - 1
-    pos = row * 8 + col
-    return 1 << (63 - pos)
-
-
-def main():
-    board = BitBoard.initial()
-    black_to_move = True
-    while True:
-        print(board)
-        player = 'Black' if black_to_move else 'White'
-        legal = board.legal_moves(board.black if black_to_move else board.white,
-                                  board.white if black_to_move else board.black)
-        if legal == 0:
-            print(f"{player} has no moves. Pass.")
-            black_to_move = not black_to_move
-            if board.legal_moves(board.black if black_to_move else board.white,
-                                 board.white if black_to_move else board.black) == 0:
-                print("No moves for both players. Game over.")
-                break
-            continue
-        move_str = input(f"{player} move (e.g., d3) or 'q' to quit: ")
-        if move_str.lower() == 'q':
-            break
-        try:
-            move = parse_move(move_str)
-            board = board.apply_move(move, black_to_move)
-            black_to_move = not black_to_move
-        except ValueError as e:
-            print(f"Invalid move format: {e}. Please enter a valid move (e.g., d3).")
-        except InvalidMoveError as e:
-            print(f"Illegal move: {e}. Please try again.")
-
-    b_count = bin(board.black).count('1')
-    w_count = bin(board.white).count('1')
-    print(f"Final score - Black: {b_count}, White: {w_count}")
-
-
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
